+++ cli/commands/watch/error-checker.ts
@@ -1,0 +1,87 @@
+import {execFile} from 'node:child_process';
+import {promisify} from 'node:util';
+import os from 'node:os';
+import chalk from 'chalk';
+
+import {getMocPath} from '../../helpers/get-moc-path.js';
+import {getRootDir} from '../../mops.js';
+import {sources} from '../sources.js';
+import {parallel} from '../../parallel.js';
+import {globMoFiles} from './globMoFiles.js';
+
+export class ErrorChecker {
+	verbose = false;
+	canisters : Record<string, string> = {};
+	status : 'pending' | 'running' | 'error' | 'success' = 'pending';
+	errors : string[] = [];
+
+	constructor({verbose, canisters} : {verbose : boolean, canisters : Record<string, string>}) {
+		this.verbose = verbose;
+		this.canisters = canisters;
+	}
+
+	reset() {
+		this.status = 'pending';
+		this.errors = [];
+	}
+
+	async run(onProgress : () => void) {
+		this.reset();
+
+		this.status = 'running';
+		onProgress();
+
+		let rootDir = getRootDir();
+		let mocPath = getMocPath();
+		let deps = await sources({cwd: rootDir});
+
+		let paths = [...Object.values(this.canisters)];
+		if (!paths.length) {
+			paths = globMoFiles(rootDir);
+		}
+
+		await parallel(os.cpus().length, paths, async (file) => {
+			try {
+				await promisify(execFile)(mocPath, ['--check', ...deps.flatMap(x => x.split(' ')), file], {cwd: rootDir});
+			}
+			catch (error : any) {
+				error.message.split('\n').forEach((line : string) => {
+					if (line.match(/: \w+ error \[/)) {
+						// better formatting
+						let str = line
+							.replace(/: (\w+ error) \[/, (_, m1) => `: ${chalk.red(m1)} [`)
+							.replace(/unbound type (\w+)/, `unbound type ${chalk.bold('$1')}`)
+							.replace(/unbound variable (\w+)/, `unbound variable ${chalk.bold('$1')}`)
+							.trim();
+						this.errors.push(str);
+					}
+					else if (line.startsWith('  ') && this.errors.length) {
+						this.errors[this.errors.length - 1] += '\n  ' + line;
+					}
+					else {
+						// console.log('UNKNOWN ERROR', line);
+					}
+				});
+				onProgress();
+			}
+		});
+
+		this.status = this.errors.length ? 'error' : 'success';
+		onProgress();
+	}
+
+	getOutput() : string {
+		if (this.status === 'pending') {
+			return `Errors: ${chalk.gray('(pending)')}`;
+		}
+		if (this.status === 'running') {
+			return `Errors: ${chalk.gray('(running)')}`;
+		}
+		let output = '';
+		output += `Errors: ${chalk.bold[this.errors.length ? 'redBright' : 'green'](this.errors.length)}`;
+		if (this.verbose && this.errors.length) {
+			output += `\n  ${this.errors.join('\n  ')}`;
+		}
+		return output;
+	}
+}